app/pyproject.toml
[project]
name = "list_cutter"
version = "0.1.0"
description = ""
authors = [
    {name = "Emily Flambe"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "whitenoise (>=6.8.2,<7.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "django-adminplus (>=0.6,<0.7)",
    "requests (>=2.32.3,<3.0.0)",
    "django-cors-headers (>=4.6.0,<5.0.0)",
    "black (>=24.10.0,<25.0.0)",
    "django-vite (>=3.0.6,<4.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "rest-framework-simplejwt (>=0.0.2,<0.0.3)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "django-extensions (>=3.2.3,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
